Create EnemyObjectSpawning.cs
This script take cares of the random spawning of the incoming objects (sprites). The choices available in terms of shapes are square and circle. On the other, the direction (left / right) is also totally randomized. A specific delay has also been provided to account for the periodicity of the spawning.